auto: issue #127 Split slurm_tool.py (1,277 lines)
